swap logic generalize (#616)
## 1. Overview

- Generalized and cleaned up swap logic
- Updated attributes and moved CoinList to generic helpers
- Added Rust dependencies and updated README
- Refactored CW DEX implementation after refactoring and monorepoing it
- Updated optimizer to v16 and merged changes from upstream feature
branch
- Refactored tests
- Solved some TODOs
- Range middleware adjustments accordingly to cl-vault ModifyRange msg


## 2. Implementation details

<!-- Describe the implementation (highlights only) as well as design
rationale. -->

## 3. How to test/use

<!-- How can people test/use this? -->

## 4. Checklist

<!-- Checklist for PR author(s). -->

- [ ] Does the Readme need to be updated?

## 5. Limitations (optional)

<!-- Describe any limitation of the capabilities listed in the Overview
section. -->

## 6. Future Work (optional)

<!-- Describe follow up work, if any. -->

---------
