+++ README.md
@@ -36,3 +36,11 @@ make install
 ## Attributions
 
 x/qtransfer, x/epochs and x/tokenfactory module are utilised from the osmosis x/ibc_hooks, x/epochs and x/tokenfactory module.
+
+
+## Dependencies
+### Rust
+In order to run test-tube the following dependencies are required:
+* `sudo apt update && sudo apt install -y build-essential pkg-config libssl-dev curl clang libclang-dev`
+* go1.21 ([see here](https://go.dev/doc/install))
+* libwasmvm ([see here](https://github.com/CosmWasm/wasmvm) -- !Instructions don't cover installation, copy the files to your desired install location or add the subfolder `wasmvm/internal/api` to your library paths)
