+++ smart-contracts/contracts/cl-vault/src/helpers/coinlist.rs
@@ -1,58 +1,10 @@
 use cosmwasm_schema::cw_serde;
-use cosmwasm_std::{
-    coin, to_json_binary, Attribute, BankMsg, Coin, CosmosMsg, Decimal, Deps, Env, Fraction,
-    Response, SubMsg, Uint128,
-};
-use osmosis_std::types::{
-    cosmos::base::v1beta1::Coin as OsmoCoin,
-    osmosis::concentratedliquidity::v1beta1::{MsgCollectIncentives, MsgCollectSpreadRewards},
-};
+use cosmwasm_std::{coin, Attribute, BankMsg, Coin, CosmosMsg, Decimal, Fraction, Uint128};
+use osmosis_std::types::cosmos::base::v1beta1::Coin as OsmoCoin;
 
-use crate::{helpers::sort_tokens, msg::ExecuteMsg, state::POSITION, ContractError};
+use crate::ContractError;
 
-/// Prepends a callback to the contract to claim any rewards, used to
-/// enforce the claiming of rewards before any action that might
-/// cause Osmosis to collect rewards anyway, such as fully withdrawing a position
-/// or adding funds into a position
-pub fn prepend_claim_msg(env: &Env, response: Response) -> Result<Response, ContractError> {
-    let claim_msg = CosmosMsg::Wasm(cosmwasm_std::WasmMsg::Execute {
-        contract_addr: env.contract.address.to_string(),
-        msg: to_json_binary(&ExecuteMsg::VaultExtension(
-            crate::msg::ExtensionExecuteMsg::CollectRewards {},
-        ))?,
-        funds: vec![],
-    });
-
-    Ok(prepend_msg(response, SubMsg::new(claim_msg)))
-}
-
-/// Prepend a msg to the start of the messages in a response
-fn prepend_msg(mut response: Response, msg: SubMsg) -> Response {
-    response.messages.splice(0..0, vec![msg]);
-    response
-}
-
-pub fn get_collect_incentives_msg(
-    deps: Deps,
-    env: Env,
-) -> Result<MsgCollectIncentives, ContractError> {
-    let position = POSITION.load(deps.storage)?;
-    Ok(MsgCollectIncentives {
-        position_ids: vec![position.position_id],
-        sender: env.contract.address.into(),
-    })
-}
-
-pub fn get_collect_spread_rewards_msg(
-    deps: Deps,
-    env: Env,
-) -> Result<MsgCollectSpreadRewards, ContractError> {
-    let position = POSITION.load(deps.storage)?;
-    Ok(MsgCollectSpreadRewards {
-        position_ids: vec![position.position_id],
-        sender: env.contract.address.into(),
-    })
-}
+use super::generic::sort_tokens;
 
 /// COIN LIST
 
@@ -103,7 +55,6 @@ impl CoinList {
         Ok(())
     }
 
-    // TODO: Cant we get Coins from a coinlist and use above function?
     pub fn update_rewards_coin_list(&mut self, rewards: CoinList) -> Result<(), ContractError> {
         let parsed_rewards: Result<Vec<Coin>, ContractError> = rewards
             .coins()
@@ -214,7 +165,7 @@ impl CoinList {
 
 #[cfg(test)]
 mod tests {
-    use cosmwasm_std::{testing::mock_env, Uint128};
+    use cosmwasm_std::Uint128;
 
     use super::*;
 
@@ -270,59 +221,6 @@ mod tests {
             positive_coins,
             vec![coin(1000, "uosmo"), coin(3000, "uqsr"),]
         );
-    }
-
-    #[test]
-    fn test_prepend_msg_with_empty_response() {
-        let response = Response::default();
-        let msg = CosmosMsg::Bank(BankMsg::Burn {
-            amount: vec![coin(100, "stake")],
-        });
-
-        let updated_response = prepend_msg(response, SubMsg::new(msg.clone()));
-        assert_eq!(updated_response.messages.len(), 1);
-        assert_eq!(updated_response.messages[0].msg, msg);
-    }
-
-    #[test]
-    fn test_prepend_msg_with_non_empty_response() {
-        let existing_msg = CosmosMsg::Bank(BankMsg::Send {
-            to_address: "bob".to_string(),
-            amount: vec![coin(100, "stake")],
-        });
-        let new_msg = CosmosMsg::Bank(BankMsg::Burn {
-            amount: vec![coin(100, "stake")],
-        });
-
-        let response = Response::new().add_message(existing_msg.clone());
-
-        let updated_response = prepend_msg(response.clone(), SubMsg::new(new_msg.clone()));
-        assert_eq!(updated_response.messages.len(), 2);
-        assert_eq!(updated_response.messages[0].msg, new_msg);
-        assert_eq!(updated_response.messages[1].msg, existing_msg);
-    }
-
-    #[test]
-    fn test_prepend_claim_msg_normal_operation() {
-        let env = mock_env();
-        let msg = CosmosMsg::Bank(BankMsg::Burn {
-            amount: vec![coin(100, "stake")],
-        });
-        let response = Response::new().add_message(msg.clone());
-
-        let claim_msg = CosmosMsg::Wasm(cosmwasm_std::WasmMsg::Execute {
-            contract_addr: env.contract.address.to_string(),
-            msg: to_json_binary(&ExecuteMsg::VaultExtension(
-                crate::msg::ExtensionExecuteMsg::CollectRewards {},
-            ))
-            .unwrap(),
-            funds: vec![],
-        });
-
-        let updated_response = prepend_claim_msg(&env, response).unwrap();
-        assert_eq!(updated_response.messages.len(), 2);
-        assert_eq!(updated_response.messages[0].msg, claim_msg);
-        assert_eq!(updated_response.messages[1].msg, msg);
     }
 
     #[test]
