--- smart-contracts/contracts/cl-vault/src/rewards/mod.rs
@@ -1,5 +1,0 @@
-mod distribution;
-mod helpers;
-
-pub use distribution::*;
-pub use helpers::*;
