+++ smart-contracts/contracts/cl-vault/src/vault/mod.rs
@@ -2,6 +2,7 @@ pub mod admin;
 pub mod any_deposit;
 pub mod autocompound;
 pub mod concentrated_liquidity;
+pub mod distribution;
 pub mod exact_deposit;
 pub mod merge;
 pub mod range;
